Began work on font name adjustment
Implemented a font-file parser using opentype.js that parses the head and name tables from the font as raw data. It then adjusts the names in the Name field, the name table checksum, updates the table directory offsets, and finally updates the entire font checksum.

As of this commit it works for many fonts, but the checksum is incorrect for many others (eg. https://fonts.adobe.com/fonts/altivo)