Added CORS Proxy cycler.
The new `getNetworkResource` will request a network resource using every CORS proxy in a list until it either succeeds or all proxies fail.
If one succeeds, it will save the index of this proxy in the browser's localstorage so it can automatically use this one the next time the user makes a network request.